# docs/swagger.yaml
basePath: /
definitions:
...
  models.Age:
    properties:
      age:
        type: string
      id:
        type: integer
      posterUrl:
        type: string
    type: object
  models.AllSeries:
    properties:
      description:
        type: string
      director:
        type: string
      id:
        description: Указатель на int
        type: integer
      rating:
        type: integer
      release_year:
        type: integer
      series:
        type: integer
      title:
        type: string
      trailer_url:
        type: string
    type: object
  models.ApiError:
    properties:
      error:
        type: string
    type: object
  models.Category:
    properties:
      id:
        type: integer
      posterUrl:
        type: string
      title:
        type: string
    type: object
  models.Genre:
    properties:
      id:
        type: integer
      posterUrl:
        type: string
      title:
        type: string
    type: object
  models.Movie:
    properties:
      ages:
        items:
          $ref: '#/definitions/models.Age'
        type: array
      allseries:
        items:
          $ref: '#/definitions/models.AllSeries'
        type: array
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      description:
        type: string
      director:
        type: string
      genres:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      id:
        type: integer
      is_watched:
        type: boolean
      poster_url:
        type: string
      rating:
        type: integer
      release_year:
        type: integer
      title:
        type: string
      trailer_url:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: "OZINSHE\tAPI"
  version: "1.0"
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: List of movies
          schema:
            items:
              $ref: '#/definitions/models.Movie'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Category model
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Invalid request category
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Create category
      tags:
      - categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Category id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid category Id
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Delete category
      tags:
      - categories
    get:
      consumes:
      - application/json
      parameters:
      - description: Category id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid Movie Id
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Find by id
      tags:
      - categories
    put:
      consumes:
      - application/json
      parameters:
      - description: Category model
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Invalid Category Id
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update category
      tags:
      - categories
  /genres:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: List of genres
          schema:
            items:
              $ref: '#/definitions/models.Genre'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all genres
      tags:
      - genres
    post:
      consumes:
      - application/json
      parameters:
      - description: Genre model
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Genre'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Invalid request category
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Create genre
      tags:
      - genres
  /genres/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Genre id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/models.Genre'
        "400":
          description: Invalid genre Id
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Delete genre
      tags:
      - genres
    get:
      consumes:
      - application/json
      parameters:
      - description: Genre id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/models.Genre'
        "400":
          description: Invalid Movie Id
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Find by id
      tags:
      - genres
    put:
      consumes:
      - application/json
      parameters:
      - description: Genre model
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Genre'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Invalid Genre Id
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update genre
      tags:
      - genres
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"